Unify routes
Bus routes output by the API consist of set of points (latitude, longitude),
partitioned by route IDs. These partitions of points are ordered.

On the other hand, itineraries are presented as set of bus stops, including
their coordinates, along with the naming of the direction associated with each
bus stop.

In this commit, an attempt has been made at unifying routes and itineraries
so as to obtain, for a given line number, a circular, ordered set of points
in its path, including bus stops. This unified route will then allow modeling.